♻️ Address review: make logging inert, dedupe redaction helpers
- useCompletionEffect: the hasSessionCookie check ran before the login
  redirect inside the same try block, so a failing server action would have
  silently swallowed the redirect. It now runs after router.push, fire and
  forget, with its own error capture.
- useVerifyEffect: the rejection breadcrumbs/captureMessage ran before the
  dispatch, so a throwing logger would have blocked the error feedback.
  Dispatch now comes first.
- proxy: x-request-id is always regenerated on the main pass instead of
  trusting the client-provided header (correlation id, anti-spoofing).
- maskEmail/truncateUserId: moved to packages/core/lib/pii as the single
  implementation; apps/server logger.ts and apps/site utils/maskEmail.ts
  re-export it, removing the duplicate implementations.
