+++ apps/server/src/logger.ts
@@ -1,6 +1,9 @@
+import { maskEmail, truncateUserId } from '@nosgestesclimat/core/lib/pii'
 import winston from 'winston'
 import SentryTransport from 'winston-transport-sentry-node'
 import { config } from './config.ts'
+
+export { maskEmail, truncateUserId }
 
 const { combine, timestamp, json, errors } = winston.format
 
@@ -19,11 +22,6 @@ if (config.thirdParty.sentry.dsn) {
     })
   )
 }
-
-export const truncateUserId = (userId: unknown) =>
-  typeof userId === 'string'
-    ? `${userId.slice(0, 8)}***`
-    : '[REDACTED]'
 
 export const redactBody = <T = unknown>(body: T) => {
   if (typeof body === 'object' && !!body) {
@@ -59,22 +57,6 @@ export const redactBody = <T = unknown>(body: T) => {
   }
 
   return body
-}
-
-/**
- * Keeps a log line correlatable with a user report without storing the address:
- * `jo***@ex***.com` is enough to match an email a user gives us in support.
- */
-export const maskEmail = (email: unknown) => {
-  if (typeof email !== 'string') {
-    return '[REDACTED]'
-  }
-
-  const [local, domain] = email.split('@')
-
-  return domain
-    ? `${local.slice(0, 2)}***@${domain.slice(0, 2)}***`
-    : '[REDACTED]'
 }
 
 /**
